ARTEMIS-1009 Fixing compatibility
there was an issue with running older clients
Found by this testsuite:

https://github.com/clebertsuconic/messaging-versioning

this will fix further compatibility issues

+++ artemis-core-client/src/test/java/org/apache/activemq/artemis/message/CoreMessageTest.java
@@ -197,7 +197,7 @@ public class CoreMessageTest {
       empty2.receiveBuffer(buffer);
 
       try {
-         empty2.getBodyBuffer().readByte();
+         empty2.getBodyBuffer().readLong();
          Assert.fail("should throw exception");
       } catch (Exception expected) {
